pyproject.toml
[tool.poetry]
name = "pipeline-resource-templates"
version = "0.1.0"
description = "Code used to generate GCP resource templates."
authors = [""]

[tool.poetry.dependencies]
python = "^3.10"
fastavro = "^1.4.10"
google-cloud-pubsub = "^2.12.0"
apache-beam = "^2.37.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
